ASMJIT_OFFSET_OF should cast to int

+++ src/asmjit/build.h
@@ -238,7 +238,7 @@
 
 //! Cross-platform solution to get offset of `_Field_` in `_Struct_`.
 #define ASMJIT_OFFSET_OF(_Struct_, _Field_) \
-  ((size_t) ((const uint8_t*) &((const _Struct_*)0x1)->_Field_) - 1)
+  static_cast<int>((intptr_t) ((const uint8_t*) &((const _Struct_*)0x1)->_Field_) - 1)
 
 // ============================================================================
 // [asmjit::build - ASMJIT_ARRAY_SIZE]
